requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Werkzeug==2.2.2
PyJWT==2.6.0
